Fix linter issues and refactor slightly
