+++ Sources/TokamakCore/Views/Layout/ZStack.swift
@@ -14,30 +14,6 @@
 
 import Foundation
 
-/// An alignment in both axes.
-public struct Alignment: Equatable {
-  public var horizontal: HorizontalAlignment
-  public var vertical: VerticalAlignment
-
-  public init(
-    horizontal: HorizontalAlignment,
-    vertical: VerticalAlignment
-  ) {
-    self.horizontal = horizontal
-    self.vertical = vertical
-  }
-
-  public static let topLeading = Self(horizontal: .leading, vertical: .top)
-  public static let top = Self(horizontal: .center, vertical: .top)
-  public static let topTrailing = Self(horizontal: .trailing, vertical: .top)
-  public static let leading = Self(horizontal: .leading, vertical: .center)
-  public static let center = Self(horizontal: .center, vertical: .center)
-  public static let trailing = Self(horizontal: .trailing, vertical: .center)
-  public static let bottomLeading = Self(horizontal: .leading, vertical: .bottom)
-  public static let bottom = Self(horizontal: .center, vertical: .bottom)
-  public static let bottomTrailing = Self(horizontal: .trailing, vertical: .bottom)
-}
-
 /// A view that overlays its children, aligning them in both axes.
 ///
 ///     ZStack {
