--- Sources/TokamakDOM/index.html
@@ -1,123 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <title>Tokamak App</title>
-    <script>
-      window.Tokamak = {
-        root: undefined,
-        measureText: (text, proposedSize) => {
-          const element = document.createElement("span")
-          element.textContent = text
-          element.style.maxWidth = `${proposedSize.width}px`
-          element.style.maxHeight = `${proposedSize.height}px`
-          document.body.appendChild(element)
-          const rect = element.getBoundingClientRect()
-          const size = {
-            width: rect.width,
-            height: rect.height
-          }
-          document.body.removeChild(element)
-          return size
-        },
-        initRoot: (root) => {
-          document.body.style.margin = '0'
-          root.style.width = '100vw'
-          root.style.height = '100vh'
-          root.style.position = 'relative'
-          window.Tokamak.root = root
-        },
-        create: (element) => {
-          const result = document.createElement(element.tag)
-          for (const attribute of element.attributes) {
-            if (attribute.property)
-              result[attribute.name] = attribute.value
-            else
-              result.setAttribute(attribute.name, attribute.value)
-          }
-          if (!!element.innerHTML)
-            result.innerHTML = element.innerHTML
-          for (const [event, action] of Object.entries(element.listeners))
-            result.addEventListener(event, action)
-          result.style.position = 'absolute'
-          element.bind(result)
-          return result
-        },
-        commit: (mutations) => {
-          for (const mutation of mutations) {
-            switch (mutation.operation) {
-            case 0: // .insert
-              const parent = mutation.parent() ?? window.Tokamak.root
-              const element = window.Tokamak.create(mutation.element)
-              element.dataset.view = mutation.element.debugData.view
-              if (parent.children.length > mutation.index) {
-                parent.insertBefore(element, parent.children[mutation.index])
-              } else {
-                parent.appendChild(element)
-              }
-              break
-            case 1: // .remove
-              if (!mutation.element) {
-                console.error(`Cannot remove null from parent.`)
-                continue
-              }
-              mutation.element.remove()
-              break
-            case 2: // .replace
-              const replaceParent = mutation.parent() ?? window.Tokamak.root
-              const original = mutation.element()
-              const replacement = window.Tokamak.create(mutation.replacement)
-              replaceParent.replaceChild(replacement, original)
-              break
-            case 3: // .update
-              // If we set the .style property, we need to copy the layout styles back in after.
-              const geometry = {
-                width: mutation.previous.style.width,
-                height: mutation.previous.style.height,
-                left: mutation.previous.style.left,
-                top: mutation.previous.style.top
-              }
-              for (const attribute of mutation.updates.attributes) {
-                if (attribute.property)
-                  mutation.previous[attribute.name] = attribute.value
-                else
-                  mutation.previous.setAttribute(attribute.name, attribute.value)
-              }
-              mutation.previous.innerHTML = mutation.updates.innerHTML
-              for (const [event, action] of Object.entries(mutation.updates.listeners))
-                mutation.previous.addEventListener(event, action)
-              mutation.updates.bind(mutation.previous)
-              // Copy back the layout.
-              mutation.previous.style.position = 'absolute'
-              for (const [style, value] of Object.entries(geometry))
-                mutation.previous.style[style] = value
-              break
-            case 4: // .layout
-              const target = mutation.element()
-              target.style.position = 'absolute'
-              target.style.width = `${mutation.geometry.width}px`
-              target.style.height = `${mutation.geometry.height}px`
-              target.style.left = `${mutation.geometry.x}px`
-              target.style.top = `${mutation.geometry.y}px`
-              continue
-            default:
-              console.error(`Unhandled mutation ${mutation.operation}`, mutation)
-              continue
-            }
-          }
-        }
-      }
-    </script>
-  </head>
-  <body>
-    <!--
-      Your Tokamak app will be injected into the #root element,
-      unless a different mountpoint is specified in the renderer.
-    -->
-    <div id="root"></div>
-    <noscript>
-      You need to enable JavaScript to run this app.
-    </noscript>
-  </body>
-</html>
